refactor: cleanup to put utils and constants in separate files
Moves various constants out to a separate constants.js file, moves getVideoState to utils.js file so
that things are more consistently sorted in a way that makes sense.
